requirements.txt
Crypto
BeautifulSoup==3.2.1
lxml==3.1beta1
pyenchant==1.6.5
pydns>=2.3.6
watchdog
yql
tweepy
pygeoip
BeautifulSoup4
pycrypto
